Rename HALO in other components.
