+++ public/locales/en.json
@@ -104,7 +104,7 @@
   "earned": "Earned",
   "getBPTTokens": "To stake, get BPT Tokens here",
   "closeTxt": "Close",
-  "stakeToEarn": "Stake your Balancer (Liquidty Pool) Pool Tokens (BPT) to earn Halo!",
+  "stakeToEarn": "Stake your Balancer (Liquidty Pool) Pool Tokens (BPT) to earn Rainbow Candy (RNBW)!",
   "learnAboutStaking": "Learn about staking",
   "approve": "Approve",
   "approving": "Approving",
